feat: implement new breadcrumb component and remove legacy index.js
