+++ docs/website/src/theme/DocBreadcrumbs/Items/Home/index.tsx
@@ -1,11 +1,14 @@
-import React from 'react';
+import React, {type ReactNode} from 'react';
 import Link from '@docusaurus/Link';
 import useBaseUrl from '@docusaurus/useBaseUrl';
 import {translate} from '@docusaurus/Translate';
 import IconHome from '@theme/Icon/Home';
+
 import styles from './styles.module.css';
-export default function HomeBreadcrumbItem() {
+
+export default function HomeBreadcrumbItem(): ReactNode {
   const homeHref = useBaseUrl('/');
+
   return (
     <li className="breadcrumbs__item">
       <Link
